Delete ParserUtil test remnents
